+++ core/libs/Base/cmtkImageOperationReplace.h
@@ -86,7 +86,7 @@ public:
 	break;
       case Self::REPLACE_INF_NAN:
 #pragma omp parallel for
-	for ( size_t i = 0; i < volumeData.GetDataSize(); ++i )
+	for ( int i = 0; i < static_cast<int>( volumeData.GetDataSize() ); ++i )
 	  {
 	  cmtk::Types::DataItem value = 0;
 	  if ( volumeData.Get( value, i ) )
